app/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-cosmos==4.3.1
azure-identity==1.13.0b3
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Flask==2.2.5
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
openai==0.27.3
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tqdm==4.65.0
typing_extensions==4.6.1
urllib3==1.26.16
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
